Moved Alert function definition. Changed Alert to take the element that will display the alert as a parameter.
